moved firebase.ts file
