--- src/services/firebase/analytics/analytics.ts
@@ -1,14 +1,0 @@
-import { analytics } from "../firebase";
-import { logEvent } from "firebase/analytics";
-
-/**
- * Logs an event to Google Analytics
- *
- * @param eventName
- * @param eventParams
- */
-const logAnalyticsEvent = (eventName: string, eventParams: any) => {
-  logEvent(analytics, eventName, eventParams);
-};
-
-export { logAnalyticsEvent };
