+++ src/services/firestore/graph/short-urls.ts
@@ -1,5 +1,5 @@
 import { collection, doc, getDoc, setDoc } from "firebase/firestore";
-import { db } from "../firebase";
+import { db } from "../../firebase";
 
 /**
  * Get the original url from the database
