+++ src/services/firestore/user/search-history.ts
@@ -8,7 +8,7 @@ import {
   setDoc,
   where,
 } from "firebase/firestore";
-import { db } from "../firebase";
+import { db } from "../../firebase";
 
 export interface AddressHistoryAPIObject {
   address: string;
